+++ Magento/Mtf/Client/BrowserInterface.php
@@ -60,6 +60,13 @@ interface BrowserInterface
     public function switchToFrame(Locator $locator = null);
 
     /**
+     * Open new tab/window in Browser and switch to it.
+     *
+     * @return void
+     */
+    public function openNewWindow();
+
+    /**
      * Close the current window or specified one.
      *
      * @param string|null $handle [optional]