+++ Magento/Mtf/Client/Driver/Selenium/Driver.php
@@ -785,6 +785,16 @@ class Driver implements DriverInterface
     }
 
     /**
+     * Open new tab/window in Browser.
+     *
+     * @return void
+     */
+    public function openWindow()
+    {
+        $this->driver->execute(['script' => 'window.open()', 'args' => []]);
+    }
+
+    /**
      * Close the current window or specified one.
      *
      * @param string|null $handle [optional]