+++ Magento/Mtf/Client/DriverInterface.php
@@ -205,6 +205,13 @@ interface DriverInterface
     public function switchToFrame(Locator $locator);
 
     /**
+     * Open new tab in Browser.
+     *
+     * @return void
+     */
+    public function openWindow();
+
+    /**
      * Close the current window or specified one.
      *
      * @param string|null $handle [optional]