+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added an ability to open new tab/window
+
 1.0.0-rc.49
 ===========
 * Added isPresent method to the ElementInterface